sudoku-back/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaggle==1.5.13
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
requests==2.28.2
six==1.16.0
smmap==5.0.0
text-unidecode==1.3
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
